Add inline-alloc support for conservative scanning
Conservative stack scanning relied on the allocation call for each new
object to add its entry in the parent page's object start bitmap. This
precluded inline allocation and allocation folding because no such
bottleneck could be guaranteed.

This updates the ConservativeStackVisitor to no longer rely on such
constraint. The object start bitmap is now assumed to be imprecise: if a
pointer doesn't correspond to a bitmap entry, the stack visitor will
fallback to the slower method of iterating page objects. In other words,
the object start bitmap can be thought of as an optimization on page
iteration, and *not* necessary for correctness. This means that it can
be used with inline allocation and allocation folding.

The object start bitmap is reconstructed during sweeping, where the
position of live objects is known. This means that visiting a root to an
inline allocated object will take the slow path once during the next
garbage collection, but be precise thereafter.

Bug: v8:10614
Signed-off-by: Jake Hughes <jh@jakehughes.uk>